completa a associaçao entre os dois tipos de utilizador

+++ src/main/java/dev/cleaningservice/dao/UserInfoDAO.java
@@ -5,7 +5,7 @@ import dev.cleaningservice.entity.UserInfo;
 
 public interface UserInfoDAO {
 
-    public UserInfo findUserInfoById(int id);
+    public UserInfo findUserInfoById(Long id);
 
     public void save(UserInfo userInfo);
